requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
djangorestframework-simplejwt>=4.6.0,<4.7.0

# tensorflow>=2.2.0,<2.3.0
flake8>=3.8.4,<3.9.0
# psycopg2>=2.8.6,<2.9.0

drf-yasg>=1.20.0,<1.21.0
